backend/requirements.txt
sentence-transformers==2.2.2
chromadb==0.4.22
numpy
python-dotenv